Extract enums for shippings
